+++ src/TourWindCaster.java
@@ -19,7 +19,7 @@ public class TourWindCaster extends Tours {
     }
 
     @Override
-    public void draw(double x, double y, int rayon) {
+    public void draw(double x, double y, double rayon) {
         StdDraw.setPenColor(StdDraw.WHITE);
         StdDraw.filledCircle(x,y,rayon);
     }
